create failing timeout test

+++ tchannel/schemes/raw.py
@@ -73,7 +73,7 @@ class RawArgScheme(object):
             A raw body to provide to the endpoint.
         :param string headers:
             A raw headers block to provide to the endpoint.
-        :param int timeout:
+        :param float timeout:
             How long to wait (in seconds) before raising a ``TimeoutError`` -
             this defaults to ``tchannel.glossary.DEFAULT_TIMEOUT``.
         :param string retry_on:
